fix build for 64bit systems
* Makefile: $ARCH is unused
* signelf.cpp: make nSigLen conform RSA_sign()'s signature

+++ signelf.cpp
@@ -102,7 +102,7 @@ namespace signelf
 		// if we could read the key
 		if(pKey)
 		{
-			size_t nSigLen;
+			unsigned int nSigLen;
 			arRetval.resize(RSA_size(pKey));
 
 			// sign the hash and save it to arRetval
